Poketimmer/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
